// editor/editor-plugin-custom-autoformat/src/types.ts
import { type AutoformattingProvider } from '@atlaskit/editor-common/provider-factory';
import type { NextEditorPlugin } from '@atlaskit/editor-common/types';
import type { Node as ProsemirrorNode } from '@atlaskit/editor-prosemirror/model';
export type {
	AutoformatHandler,
	AutoformatReplacement,
	AutoformattingProvider,
	AutoformatRuleset as Ruleset,
} from '@atlaskit/editor-common/provider-factory';

export type Reducer<ActionType> = (
	state: CustomAutoformatState,
	action: ActionType,
) => CustomAutoformatState;

export type CustomAutoformatPlugin = NextEditorPlugin<
	'customAutoformat',
	{
		sharedState: CustomAutoformatPluginSharedState | undefined;
		pluginConfiguration: CustomAutoformatPluginOptions;
		actions: {
			setProvider: (provider: Promise<AutoformattingProvider>) => Promise<boolean>;
		};
	}
>;

export type CustomAutoformatPluginOptions = {
	autoformattingProvider: Promise<AutoformattingProvider> | undefined;
};

export type AutoformatCandidate = {
	start: number;
	end: number;
	match: string[];
};

export type AutoformatMatch = {
	matchString: string;
	replacement?: ProsemirrorNode;
};

export type CustomAutoformatPluginSharedState = {
	autoformattingProvider: AutoformattingProvider | undefined;
};

export type CustomAutoformatState = CustomAutoformatPluginSharedState & {
	resolving: Array<AutoformatCandidate>;
	matches: Array<AutoformatMatch>;
};

// actions

export type CustomAutoformatMatched = {
	action: 'matched';
	start: number;
	end: number;
	match: string[];
};

export type CustomAutoformatResolved = {
	action: 'resolved';
	matchString: string;
	replacement?: ProsemirrorNode;
};

export type CustomAutoformatFinish = {
	action: 'finish';
	matchString: string;
};

export type CustomAutoformatSetProvider = {
	action: 'setProvider';
	provider: AutoformattingProvider | undefined;
};

export type CustomAutoformatAction =
	| CustomAutoformatMatched
	| CustomAutoformatResolved
	| CustomAutoformatFinish
	| CustomAutoformatSetProvider;